Report test execution progress using signals
This allows us to have more incremental updates, rather than just
awaiting on the test result, and so allowing us to distinguish between
"not started" and "currently running" tests.

I'm not entirely happy with this design, but it's a good start at least.
Reimplementing a minimum subset of React feels a little weird, but I'm
not entirely sure their design works for our needs.