Remove our Lwt dependency
While it was conceptually a nice idea to have everything execute
asynchronously in parallel, it makes it a bit painful to add
compatibility with other test frameworks (no capturing of stdout, how to
handle tests which use Lwt_main.run, etc…).

Further more, as we're using signals instead of promises for test
progress, there wasn't much point really using Lwt within the ingredient
system - it's possible (albeit a teeny bit confusing) to model our
system without it.

We also remove the requirement for reporters to return a success value
or not - this was copied from tasty, but it's really not apparent of the
use-case for this, and ends up complicating any existing reporters.